move run to a separate file
